Module lu_packets::auth::server

Server-received auth messages.

Re-exports

pub use crate::general::server::GeneralMessage;

Structs

ComputerStats

Stats about the computer the client is running on.

LoginRequest

Provides username and password to authenticate the client.

OsInfo

Info about the operating system the client is running on.

ProcessorInfo

Info about the processor the client is running on.

Enums

AuthMessage

All server-received auth messages.

ClientOs

The client’s operating system.

LuMessage

All LU messages that can be received by an auth server.

Type Definitions

Message

All messages that can be received by an auth server.