Enum lu_packets::general::client::DisconnectNotify

#[repr(u32)]pub enum DisconnectNotify {
    UnknownServerError,
    WrongGameVersion(u32),
    WrongServerVersion(u32),
    ConnectionOnInvalidPort,
    DuplicateLogin,
    ServerShutdown,
    UnableToLoadMap,
    InvalidSessionKey,
    AccountNotInPendingList,
    CharacterNotFound,
    CharacterCorruption,
    Kick,
    SaveFailure,
    FreeTrialExpired,
    PlayScheduleTimeUp,
}

Notifies the client when it was actively disconnected by the server.

Trigger

Being disconnected by the server, the exact trigger depends on the variant.

Handling

Display a message to the user.

Response

None. Expect the connection to be closed shortly after, so a response won't even be possible.

Notes

You can be disconnected without receiving this packet, for example when your connection is lost. The server is also not obligated to send this packet and may disconnect you without doing so.

Variants

UnknownServerError

Unspecified disconnect reason.

WrongGameVersion(u32)

The client's network_version did not match the server's network_version. The message contains the server network version number.

WrongServerVersion(u32)

Unused for client-server.

ConnectionOnInvalidPort

Connection attempt on invalid port, server emulators probably won't send this as they usually won't have server-server communication using LU's protocol.

DuplicateLogin

There was another login with your account and your session has been closed in favor of the new login.

ServerShutdown

The server is shutting down.

UnableToLoadMap

No server hosting this map is available.

InvalidSessionKey

The provided ClientValidation::session_key is incorrect.

AccountNotInPendingList

Server did not expect a ClientValidation at this time.

CharacterNotFound

The provided CharacterLoginRequest::char_id was not a valid character ID of this account.

CharacterCorruption

The character seems to be corrupted in the database.

Kick

You were kicked from the server.

SaveFailure

Error saving or loading progress.

FreeTrialExpired

The account's time-limited free trial expired, unused.

PlayScheduleTimeUp

The parental controls for this account prevent it from further play.