Module lu_packets::world::client

Client-received world messages.

Structs

AddFriendRequest
Informs the client that another player has asked them to be their friend.

AddFriendResponse

BlueprintLoadItemResponse

CharListChar
A character from the CharacterListResponse message.

CharacterDeleteResponse
Reports the result of a character delete request.

CharacterListResponse
Provides the list of characters of the client's account.

ChatModerationString

CreateCharacter

FriendState

FriendUpdateNotify

IgnoreState

LoadStaticZone
Tells the client to load a zone.

MinimumChatModeResponse

MinimumChatModeResponsePrivate

ModerationSpan

TeamInvite
Informs the client that another player has asked them to be their friend.

TransferToWorld
Tells the client to open a connection to another server instance.

UpdateFreeTrialStatus
Notifies the client that its free trial status has changed.

Enums

AddFriendResponseType

CharacterCreateResponse
Reports the result of a character create request.

ClientMessage
All client-received world messages.

FriendUpdateType

GetFriendsListResponse

GetIgnoreListResponse

InstanceType

LuMessage
All client-received LU messages from a world server.

Type Definitions

Message
All messages that can be received by a client from a world server.