Module lu_packets::world::server

Server-received world messages.

Re-exports

pub use crate::general::server::GeneralMessage;

Modules

mail

Structs

CharacterCreateRequest

Requests a new character to be created.

CharacterDeleteRequest

Requests a character to be deleted.

CharacterLoginRequest

Indicates that the player has chosen a character to play with.

ClientValidation

Provides session info for authentication.

GeneralChatMessage

LevelLoadComplete

Reports to the server that client-side loading has finished.

PositionUpdate

StringCheck

Asks the server whether a string the player entered is acceptable.

Top5IssuesRequest

UgcDownloadFailed

Enums

Language

LuMessage

All LU messages that can be received by a world server.

RouteMessage

UgcResType

WorldMessage

All server-received world messages.

Type Definitions

Message

All messages that can be received by a world server.